--- a/blobstore/util/retry/retry.go
+++ b/blobstore/util/retry/retry.go
@@ -15,6 +15,7 @@
 package retry
 
 import (
+	"context"
 	"errors"
 	"time"
 )
@@ -28,20 +29,49 @@ var (
 
 // Retryer is an interface retry on a specific function.
 type Retryer interface {
+	// Reset the delay as beginning.
+	Reset()
 	// On performs a retry on function, until it doesn't return any error.
 	On(func() error) error
+	// OnContext On or context done.
+	OnContext(context.Context, func() error) error
 	// RuptOn performs a retry on function, until it doesn't return any error or interrupt.
 	RuptOn(func() (bool, error)) error
+	// RuptOnContext RuptOn or context done.
+	RuptOnContext(context.Context, func() (bool, error)) error
 }
 
+// retry non-thread-safer implements Retryer
 type retry struct {
 	attempts  int
+	reset     func()
 	nextDelay func() uint32
+}
+
+type internalContext struct{}
+
+func (internalContext) Deadline() (deadline time.Time, ok bool) { return }
+func (internalContext) Done() <-chan struct{}                   { return nil }
+func (internalContext) Err() error                              { return nil }
+func (internalContext) Value(key interface{}) interface{}       { return nil }
+
+func (r *retry) Reset() {
+	if r.reset != nil {
+		r.reset()
+	}
 }
 
 // On implements Retryer.On.
 func (r *retry) On(caller func() error) error {
+	return r.OnContext(internalContext{}, caller)
+}
+
+// OnContext implements Retryer.OnContext.
+func (r *retry) OnContext(ctx context.Context, caller func() error) error {
 	var lastErr error
+	var timer *time.Timer
+	var ctxDone bool
+
 	attempt := 1
 	for attempt <= r.attempts {
 		if lastErr = caller(); lastErr == nil {
@@ -52,15 +82,29 @@ func (r *retry) On(caller func() error) error {
 		if attempt >= r.attempts {
 			break
 		}
-		time.Sleep(time.Duration(r.nextDelay()) * time.Millisecond)
 		attempt++
+
+		if timer, ctxDone = r.waitContext(ctx, timer); ctxDone {
+			break
+		}
+	}
+	if timer != nil {
+		timer.Stop()
 	}
 	return lastErr
 }
 
 // RuptOn implements Retryer.RuptOn.
 func (r *retry) RuptOn(caller func() (bool, error)) error {
+	return r.RuptOnContext(internalContext{}, caller)
+}
+
+// RuptOnContext implements Retryer.RuptOnContext.
+func (r *retry) RuptOnContext(ctx context.Context, caller func() (bool, error)) error {
 	var lastErr error
+	var timer *time.Timer
+	var ctxDone bool
+
 	attempt := 1
 	for attempt <= r.attempts {
 		interrupted, err := caller()
@@ -79,10 +123,52 @@ func (r *retry) RuptOn(caller func() (bool, error)) error {
 		if attempt >= r.attempts {
 			break
 		}
-		time.Sleep(time.Duration(r.nextDelay()) * time.Millisecond)
 		attempt++
+
+		if timer, ctxDone = r.waitContext(ctx, timer); ctxDone {
+			break
+		}
+	}
+	if timer != nil {
+		timer.Stop()
 	}
 	return lastErr
+}
+
+func (r *retry) waitContext(ctx context.Context, timer *time.Timer) (*time.Timer, bool) {
+	var wait <-chan time.Time
+
+	next := time.Duration(r.nextDelay()) * time.Millisecond
+	if _, internal := ctx.(internalContext); internal {
+		if next > 0 {
+			time.Sleep(next)
+		}
+		return timer, false
+	}
+
+	if next > 0 {
+		if timer == nil {
+			timer = time.NewTimer(next)
+		} else {
+			timer.Reset(next)
+		}
+		wait = timer.C
+	}
+
+	if wait != nil {
+		select {
+		case <-wait:
+			return timer, false
+		case <-ctx.Done():
+			return timer, true
+		}
+	}
+	select {
+	case <-ctx.Done():
+		return timer, true
+	default:
+		return timer, false
+	}
 }
 
 // Timed returns a retry with fixed interval delay.
@@ -100,6 +186,9 @@ func ExponentialBackoff(attempts int, expDelay uint32) Retryer {
 	next := expDelay
 	return &retry{
 		attempts: attempts,
+		reset: func() {
+			next = expDelay
+		},
 		nextDelay: func() uint32 {
 			r := next
 			next += expDelay