--- a/docker/run_docker.sh
+++ b/docker/run_docker.sh
@@ -30,8 +30,10 @@ EOF
     exit 0
 }
 
-#compose="docker-compose --env-file ${RootPath}/docker/run_docker.env -f ${RootPath}/docker/docker-compose-hy.yml"
-compose="docker compose --env-file ${RootPath}/docker/run_docker.env -f ${RootPath}/docker/docker-compose.yml"
+cmdcompose="docker-compose"
+docker compose version > /dev/null 2>&1 && cmdcompose="docker compose"
+#compose="${cmdcompose} --env-file ${RootPath}/docker/run_docker.env -f ${RootPath}/docker/docker-compose-hy.yml"
+compose="${cmdcompose} --env-file ${RootPath}/docker/run_docker.env -f ${RootPath}/docker/docker-compose.yml"
 
 has_go() {
     if command -v go &> /dev/null